add S3 bucket for Airflow logs and RDS database for Airflow metadata
The Airflow infrastructure currently stores logs and metadata locally
on the EC2 instance. It is more robust to store logs in a dedicated
S3 bucket and the metadata in a dedicated database (because then we
can use the LocalExecutor for Airflow, and persist state if the Airflow
instance is rebuilt).

Here we create an S3 bucket for logs and setup an RDS instance for
Airflow metadata.