Add scale policies on webserver
We add scale in and out policies on the webserver autoscaling group:
scale out by 1 instance when CPU utilization stays above 80% for 5
minutes and scale in by 1 instance when CPU utilizations stays below 20%
for 15 minutes. We provide variables to adjust these parameters.

We also add a small update to Airflow user data to calculate deployment
time.